+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-## Unreleased
+## 0.12.0 (2020-07-09)
+- Silence warnings in Ruby 2.7
 - Ensure response is passed to `ActiveSupport::Notifications` on error
 - Ensure `adapter.last_response` is set on error response